+++ front/src/App.js
@@ -32,9 +32,9 @@ import CodingtestQna from './pages/codingtest/CodingtestQna';
 import CodingtestQnaView from './pages/codingtest/CodingtestQnaView';
 import CodingtestQnaWrite from './pages/codingtest/CodingtestQnaWrite';
 import CodingtestBestCode from './pages/codingtest/CodingtestBestCode';
-import CommunityList from './pages/community/CommunityList';
-import CommunityView from './pages/community/CommunityView';
-import CommunityWrite from './pages/community/CommunityWrite';
+import CommunityList from './pages/community/List/CommunityList';
+import CommunityView from './pages/community/View/CommunityView';
+import CommunityWrite from './pages/community/Write/CommunityWrite';
 import CompanyApply from './pages/company/CompanyApply';
 import CompanyBookmark from './pages/company/CompanyBookmark';
 import FindByEmail from './pages/member/FindByEmail';
