+++ front/src/pages/community/List/Paging.js
@@ -1,0 +1,27 @@
+import React, { useState } from 'react';
+import './../../../css/community/list.css';
+import Pagination from 'react-bootstrap/Pagination';
+
+function Paging() {
+    return (
+        <Pagination className={'justify-content-center'} style={{ marginTop: '10px', marginBottom: '10px' }}>
+            <Pagination.First />
+            <Pagination.Prev />
+            <Pagination.Item>{1}</Pagination.Item>
+            <Pagination.Ellipsis />
+
+            <Pagination.Item>{10}</Pagination.Item>
+            <Pagination.Item>{11}</Pagination.Item>
+            <Pagination.Item active>{12}</Pagination.Item>
+            <Pagination.Item>{13}</Pagination.Item>
+            <Pagination.Item disabled>{14}</Pagination.Item>
+
+            <Pagination.Ellipsis />
+            <Pagination.Item>{20}</Pagination.Item>
+            <Pagination.Next />
+            <Pagination.Last />
+        </Pagination>
+    );
+}
+
+export default Paging;
