+++ front/src/pages/community/View/Article/ArticleContent.js
@@ -1,0 +1,58 @@
+import React from 'react';
+import '../../../../css/community/view.css';
+import { Col, Container, Row } from 'reactstrap';
+import { useNavigate } from 'react-router-dom';
+import UserPopup from "../../../../components/Lecture/UserPopup";
+
+function ArticleContent(){
+    return(<>
+        <div className={'article_viewer'}>
+            <div>
+                <div className="content Viewer">
+                    <div className="se-viewer se-theme-default" lang={'ko-KR'}>
+                        {/*SE_DOC_HEADER_START*/}
+                        {/*CONTENST_HEADER*/}
+                        {/*SE_DOC_HEADER_END*/}
+                        <div className={'se-main-container'}>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                            <p>내용입니다0000000000000</p>
+                        </div>
+                    </div>
+                </div>
+                <div className="AttachFileIssueLayer" style={{ display: 'none' }}>
+                    {/*공백*/}
+                </div>
+            </div>
+        </div></>)
+}
+
+export default ArticleContent;
